Added reusable authentication service to check user identity and ensure DRY principle is applied

+++ controllers/lists_controller.py
@@ -1,28 +1,38 @@
 from models.List import List
 from models.Task import Task
+from models.User import User
 from main import db
 from schemas.ListSchema import list_schema, lists_schema
 from schemas.TaskSchema import task_schema, tasks_schema
-from flask import Blueprint, request, jsonify, render_template
-from flask_jwt_extended import jwt_required
+from flask import Blueprint, request, jsonify, render_template, abort
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from services.auth_service import verify_user
 
 lists = Blueprint('lists', __name__, url_prefix="/lists")
 
 @lists.route("/", methods=["GET"])
 def list_index():
     lists = List.query.all()
-    # return jsonify(lists_schema.dump(lists))
-    return render_template("lists_index.html", lists = lists)
+    return jsonify(lists_schema.dump(lists))
+    #return render_template("lists_index.html", lists = lists)
 
 @lists.route("/", methods=["POST"])
 @jwt_required
-def list_create():
+@verify_user
+def list_create(user=None):
+
+    list = List.query.filter_by(user_id=user.id).first()
+
+    if not list:
+        return abort(400, description= "Not authorized, you need to create a list first")
+
     list_fields = list_schema.load(request.json)
 
     # create a new List object, with the data received in the request
     new_list = List()
     new_list.name = list_fields["name"]
     new_list.description = list_fields["description"]
+    new_list.user_id = user.id
 
     #add a new list to the db
     db.session.add(new_list)
@@ -46,8 +56,19 @@ def list_tasks_show(id):
 
 @lists.route("/<int:id>", methods=["DELETE"])
 @jwt_required
-def list_delete(id):
-    list = List.query.get(id)
+@verify_user
+def list_delete(id, user=None):
+    # user_id = get_jwt_identity()
+    # user = User.query.get(user_id)
+
+    # if not user:
+    #     return abort(401, decription="Invalid user")
+    
+    #list = List.query.get(id)
+    list = List.query.filter_by(id=id, user_id=user.id).first()
+    if not list:
+        return abort(400, description="Not authorized to delete other people's lists")
+        
     db.session.delete(list)
     db.session.commit()
     return jsonify(list_schema.dump(list))
